+++ HISTORY.rst
@@ -2,7 +2,7 @@
 History
 =======
 
-1.1.0 (unreleased)
+1.1.0 (2026-06-24)
 ------------------
 
 * Added discovery configuration models and management APIs.